requirements.txt
Pillow==5.1.0
scikit-learn==0.19.1
Flask==1.0.2
requests==2.18.4
scipy==1.1.0
numpy==1.14.3
opencv-python==4.1.0.25
